Update test_pipeline.py to upstream redis-py version. Moved all cluster unique tests and overwrites into test_pipeline_cluster.py
